Refactor: add function parameters to uBPF programs in a flexible way
For this, getters and setters have been added to another file. This fixed a problem where 64-bit addresses were converted to 32-bits and back to 64-bits, leading to wrong addresses being used by the plugins.
When running a uBPF program, the tcp_pcb and the different arguments are merged into a struct named tcp_ubpf_cnx_t.
The clang and llc versions used for compiling bpf programs changed from v6 to v7 in this commit.