Add simple thin-stream, no tests done
In this commit, the thin-stream is used only to retransmit directly a packet if a dupACK was received. The congestion control is not affected yet by this function. Also, the thin-stream implementation is simple: it does not take into account the fact that the connection might be in the slow start process and the number of packets is considered to be the number of bytes unacked divided by the current MSS. In a real use of the plugin, those things should be changed.
Tests were not done yet as this is harder to test. An idea could be to generate loss of packets with netem so that dupACKs happen and also to limit the sender_max window to 4 to see that the implementation is always in thin-stream. This might be the easiest, yet not perfect, way.