Change structure: cnx in pcb, option parsers organized, option writers to enable
uto, rto: add functions that enable specific TCP option writer eBPF programs in server.

uto, rto: register parsers in client. Add a function to register tcp_option_parsers that belong to an option and (possibly) an experimentation ID (ExID).

Add ebpf_options_parser_bpf_code, Add ebpf_options_parser_bpf_code_253 and Add ebpf_options_parser_bpf_code_254: arrays of strings that contain the path to an eBPF program that allows to parse a TCP option with a certain option and ExID.

tcp.c: add user timeout behaviour to LWIP. Add timeout function along with the last_ack_received_tmr variable. This variable is a way to know the last time we received a segment acknowledging new data. This is an approximation to what we should have with the User TimeOut: the time for which a segment has not been ACKed.

Add cnx to pcb instead of pcb to cnx. More suitable for LWIP since the structure is already done with the tcp_pcb's. Also allows not to create a cnx every time we call an eBPF function.

Writing options: call now a single functions in ubpf.c that will call the writer functions. At the moment, only hard-coded calls. In the future, it should be interesting to have some kind of linked list. Another interesting thing would be to have a function that determines if the option_writer should write. Indeed, at the moment, we can not allow an option_writer to write only when it wants to write. For example, if a plugin would like to write an option only under some circumstances, it can not be able to do that here. It would thus be interesting to run those functions first, to be able to modify a global ebpf_options_length that would be used when trying to get the length of the options that have been written by eBPF programs.

rto_max is now a member of cnx as it does not make any sense to have this in the tcp_pcb. This rto_max is only needed by the rto plugin, it is an extension, not a part of the tcp_pcb. At the moment, it is just a member of cnx. In the future, it might be interesting to allow a hashmap in cnx so that plugins could put and modify different values in cnx.

user_timeout is a member of the tcp_pcb as this field should already be present in the TCP implementation. It thus made sense to also include the code related to the timeout in LWIP directly. The eBPF part is more about the option (parsing and writing).

ebpf_get_options_length.c: updated but not used yet; could be interesting if updating the ebpf_options_length variable dynamically.

Should split the ubpf functions like load and start and register_functions of the other functions like set_use_rto_option().

thin-stream: commented for now, better for testing. Should provide an enable function as for the writers and parsers.