Fix tcp_accept crash, free allocated VM memory, change parsing, update UTO, change N_ARGS based on compiler
Commit afe257deece49e0e23d6023265c94491b0283316 introduced bugs as it changed the tcp_pcb structure. The bug occurred when the server kept running against a client (which required the server to call an eBPF program with an input). The bug consisted in pcb->listener->accept being replaced by 1 (the value of cnx.inputc). The reason was a collision between pcb->cnx and pcb->listener. In LWIP, to set the accept function, the pcb is cast in a tcp_pcb_listener and the accept function is thus set. When setting (pcb->cnx).inputc, it thus replaced the value of the accept function. The cnx field has been put at the end of the tcp_pcb structure.

When calling eBPF, the memory allocated for the VM was not freed, this is now the case. Also, when the VM crashes, he return value is now set to (uint64_t)-1.

N_ARGS macro: the macro now depends on the compiler according to __GNUC__ and __STDC_VERSION__.

Parsing eBPF programs: the length of the TCP options is read before calling ebpf functions. length is now an arg of the eBPF program called. Reason: easier for updating tcp_optidx, gives more power to the implementation to know what is going on. Parsing both the kind (opt) and the value (length) is not a problem, we know those will be next to each other in every TCP option.
Return value is now either ERR_OK if everything went fine, ERR_VAL if no parser was foun or if the format received was in an unexpected format, ERR_ARG if the eBPF program crashed or if the option length had a bad value (<2). The last error code causes the program to stop.

UTO option: now the UTO option parser sets user_timeout instead of rto_max. Functions get_user_timeout and set_user_timeout avec been added to the API called by eBPF (getset.c)