Fix measure_delayed_ack_perf.sh, add plot script
The measure_delayed_ack_perf.sh is meant to be sourced in order to work with the timeout functions. This allows to run only one script to do the simulations.
Added plot script. Possible to add the value of N we want to use (instead of boxplotting for N = [1, 2, 3], plot for N = [1, ..., arg]).