User Timeout: u32_t is enough + use granularity + 5min or TCP_KEEP_IDLE-1 by default
Granularity was not used before. u32_t is enough as the value is held on 15 bits and log2(60*1000) < 16 which means we need <= 31 bits to store ms, even when the field is written for minutes (granularity=1).
Also, the default value of the User Timeout should be 5 minutes. But we also know that it should be less than TCP_KEEPIDLE => we set it to the minimum, cf if-else statements