Thin_stream: measure perf with 30ms delay
Modified the measuring script to measure the performance of the thin_stream plugin when using different loss percentages with a delay of 30 ms on each link, from sender to receiver and vice-versa.
Modified also the place where the data is kept.
Plotting script now stores the result using both standard scales for X and Y or log scale for Y (duration of the transfer [ms]).