Add plugin memory + fix ebpf option TCP parser and arrays
Each plugin can now allocate memory when registering the plugin for the first time. This memory is kept thanks to the plugins_memory_t structure, which is a linkedlist whose nodes contain the plugin names and their memory (+size). Two new functions have been made available to eBPF programs: get_metadata and set_metadata, in order to read and change the values assigned to the plugin. How does one know to which plugin a pluglet belongs? The user tells which to which plugin a pluglet belongs when registering the said pluglet.
This required to include the plugin name in the tcp_ubpf_cnx_t structure and in the TCP option parser arrays.
The functions get_rto_max and set_rto_max have thus been removed as the user will now have to use the two new functions

A fix was done in ebpf_parse_tcp_option: the tcp_optidx was wrongly set after reading an option as it did not take into account the fact that the option kind was already read before calling the said function.

TCP option parser arrays for option kinds 253 and 254: the size of the arrays has changed from 256 to 65536 as the Experimental ID (ExID) is unique based on the first 16 bytes, not first 8 bytes.