Commit UTO measurement configuration+scripts v1
The configuration files are added here. The lwiperf is now using 1020 as TCP MSS as lwIP behaves strangely when using options. In fact, When adding a 4 bytes option, it is not enough to increase the SND_BUF by 4 for each packet. When using wireshark, it would be shown that the payload would never exceed 1020 bytes instead of 1024 which is the TCP MSS value.
There was maybe also another field I should have changed but simply changing the TCP MSS seems to work fine.

An outdated comment on the measure_thin_stream_perf.sh has also been removed.