Add RTO evaluation + fix problems
First, the RTO plugin has been enabled: its memory is also allocated.
The TCP_MSS for lwip is now 1016 (1024-8) as this option is 8 bytes long.
Several prints have been added (tcp.c + ebpf_should_drop_connection_rto) so that we can track the value of the RTO and compare it with wireshark data.
The script that launches the experiment has been added.