Remove useless prints RTO + restrict parsing opt buffer size + ExID RTO 134
In this commit, the prints that were used for RTO evaluation have been removed, as well as other useless prints.
The parsing option buffers (TCP option kind 253 and 254) have been restricted in size: they are 256-cells long. The reason is that this work is a proof of concept. Not everything needs to be completely implemented and supporting 256 ExIDs is enough to demonstrate that this can work. Also, 65536 cells is too much for a restrained version of TCP as lwIP. In the future, a hash map should be use instead.
The ExID for the RTO was previously stated as 0x12EF. It is changed here as 134, we should have been its first value during the experimentations. Hopefully, this does not change the results of the experiments since only the size of the parsing buffer changes really.
There is now a return value to the option pluglet register: 0 if successful, -1 otherwise.