Enable register_pluglet function
This function has been created to be able to make it easier to add pluglets, in a way that is not seen by the end user.
This function acts as a "demultiplexer", trying to find on which insertion point a function should be registered.
Accordingly, other functions that were used to call pluglets have been replaced to be more modular. A list of writer pluglets is used for the writers, an array of parser pluglets is used (as before) for parsing, and other pluglets are assigned to a single variable. This means that only one function can be registered on it.
This might appear like a limitation but we could extend it later if necessary.

Those changes have been tested manually.

The pushed configuration is the default lwIP using the UTO plugin. No other plugin is used on the pushed version. But those are functional as they have been tested.