Bug found on the uBPF version of PQUIC: infinite loop
When using a loop, if we use an unsigned int of 8 bits, it can not get to a higher value than 2^8 - 1. Thus, if looping until this variable is <= 2^8, we should have an infinite loop.
This is what is done here. The variable containing the 2^8 is dynamically defined: it is an input of the function. Without this, the eBPF verifier would have seen the problem.

To replicate this, replace the delayed ack plugin by this one, and add as an input 2^8 when calling the run_ubpf_with_args function/macro. This will stop the receiver, which will loop indefinitely.